finished users route decomposition, construct thin aggregators
